+++ packages/contracts-core/contracts/libs/memory/Snapshot.sol
@@ -161,13 +161,13 @@ library SnapshotLib {
     /// @param domain        Domain of Origin chain
     /// @param snapProof     Proof of inclusion of State Merkle Data into Snapshot Merkle Tree
     /// @param stateIndex    Index of Origin State in the Snapshot
-    function proofSnapRoot(bytes32 originRoot, uint32 domain, bytes32[] memory snapProof, uint256 stateIndex)
+    function proofSnapRoot(bytes32 originRoot, uint32 domain, bytes32[] memory snapProof, uint8 stateIndex)
         internal
         pure
         returns (bytes32)
     {
         // Index of "leftLeaf" is twice the state position in the snapshot
-        uint256 leftLeafIndex = stateIndex << 1;
+        uint256 leftLeafIndex = uint256(stateIndex) << 1;
         // Check that "leftLeaf" index fits into Snapshot Merkle Tree
         if (leftLeafIndex >= (1 << SNAPSHOT_TREE_HEIGHT)) revert IndexOutOfRange();
         // Reconstruct left sub-leaf of the Origin State: (originRoot, originDomain)
